+++ projects/net.splitcells.website.server/src/main/java/net/splitcells/website/server/Server.java
@@ -73,6 +73,7 @@ public class Server {
 
     /**
      * <p>TODO Move this code into vertx package, in order to contain {@link io.vertx} dependencies.</p>
+     * <p>TODO Always open 2 p</p>
      *
      * @param renderer renderer
      */